+++ nlp/feature_extractors/mimic_feature_extractor.py
@@ -1,0 +1,37 @@
+
+from file_utils.array_io_utils import read_array
+from nlp.stanza_utils import get_stanza_model
+from preprocessing.string_cleaning_utils import (get_punctuation,
+                                                 get_stopwords,
+                                                 remove_stopchars,
+                                                 remove_stopstrings, stem)
+from nlp.feature_extractor import FeatureExtractor
+
+class MimicFeatureExtractor(FeatureExtractor): 
+    
+    def get_features(self, text):
+        nlp = get_stanza_model(
+            package='mimic', processors={'ner': 'i2b2'}, download=False
+        )
+        doc = nlp(text)
+        result = []
+
+        # Loop through all the entities found in the text.
+        for ent in doc.entities:
+            # Construct a set of cleaned words from entity.
+            words = ent.text.split(' ')
+            words = remove_stopstrings(words, self.stopwords_)
+            words = [remove_stopchars(word, self.punctuation_) for word in words]
+            words = [stem(word) for word in words]
+            words_set = set(words)
+            # Loop through all our stored clinical features.
+            for feature in self.keyword_handler_.keywords:
+                # Check that all the words in the feature are in the set.
+                valid_feature = True
+                for word in feature:
+                    if word not in words_set:
+                        valid_feature = False
+                # If all words are found then this is likley to be a valid feature.
+                if valid_feature and len(feature) > 0:
+                    result.append(feature)
+        return result
